% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_word_doc.R
\name{add_page_break}
\alias{add_page_break}
\title{Add Page Break to Word Document}
\usage{
add_page_break(doc)
}
\arguments{
\item{doc}{An \code{officer} document object.}
}
\value{
The modified \code{officer} document object.
}
\description{
Adds a page break to a Word document.
}
\examples{
\dontrun{
doc <- table1_word_doc() |>
  implement_table1(table1_1, caption = "Table 1") |>
  add_page_break() |>
  implement_table1(table1_2, caption = "Table 2")
}

}
