% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_table1.R
\name{.create_multi_column_table1}
\alias{.create_multi_column_table1}
\title{Internal function to create a multi-column Table 1}
\usage{
.create_multi_column_table1(
  data,
  var_names,
  var_labels,
  var_subheaders,
  parsed_varlist,
  var_overrides,
  subgroups,
  include_all,
  all_label,
  digits,
  n_pct_order,
  center_fun,
  spread_fun,
  var_types,
  empty_subgroup_handling
)
}
\description{
Internal function to create a multi-column Table 1
}
\keyword{internal}
