% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_table1.R
\name{.create_table1_core}
\alias{.create_table1_core}
\title{Internal function to create a single-column Table 1}
\usage{
.create_table1_core(
  data,
  var_names,
  var_labels,
  var_subheaders,
  var_overrides,
  digits,
  n_pct_order,
  center_fun,
  spread_fun,
  var_types
)
}
\description{
Internal function to create a single-column Table 1
}
\keyword{internal}
