% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_levels.R
\name{.expand_levels_from_names}
\alias{.expand_levels_from_names}
\title{Expand Levels from Character Vector of Names}
\usage{
.expand_levels_from_names(x, label, level_names)
}
\arguments{
\item{x}{Variable vector}

\item{label}{Variable label}

\item{level_names}{Character vector of level names}
}
\value{
List of level definitions
}
\description{
Internal function to expand levels when given a character vector of level names.
}
\keyword{internal}
