% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_levels.R
\name{.expand_levels_from_values}
\alias{.expand_levels_from_values}
\title{Expand Levels from Value Mappings}
\usage{
.expand_levels_from_values(x, label, value_map)
}
\arguments{
\item{x}{Variable vector}

\item{label}{Variable label}

\item{value_map}{Named list where names are labels and values are the values to match}
}
\value{
List of level definitions
}
\description{
Internal function to expand levels when given a list mapping labels to values.
}
\keyword{internal}
