% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_yaml_varlist.R
\name{.parse_function}
\alias{.parse_function}
\title{Parse Function Name String to Function}
\usage{
.parse_function(func_name)
}
\arguments{
\item{func_name}{Character string with function name}
}
\value{
Function object
}
\description{
Internal function to convert function name strings (like "median", "mean") to actual functions.
}
\keyword{internal}
