% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_variable.R
\name{.summarize_variable_grouped}
\alias{.summarize_variable_grouped}
\title{Summarize Variable with Grouping}
\usage{
.summarize_variable_grouped(
  data,
  variable,
  label,
  var_type,
  digits,
  center_fun,
  spread_fun,
  group,
  scaling
)
}
\arguments{
\item{data}{Data frame}

\item{variable}{Character string name of the variable}

\item{label}{Label for the variable (if NULL, uses variable name)}

\item{var_type}{Type of variable (auto-detected if NULL)}

\item{digits}{Number of digits for rounding}

\item{center_fun}{Function for center statistic (mean, median, etc.)}

\item{spread_fun}{Function for spread statistic (sd, IQR, etc.)}

\item{group}{Deprecated. Not currently used.}

\item{scaling}{Scaling factor for percentages (default = 100)}
}
\value{
Data frame with group statistics and an optional difference metric
}
\description{
Internal function to summarize a variable by group.
}
\details{
TODO: This function is currently not used.
}
\keyword{internal}
