% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_word_doc.R
\name{save_table1_doc}
\alias{save_table1_doc}
\title{Save Word Document}
\usage{
save_table1_doc(doc, file)
}
\arguments{
\item{doc}{An \code{officer} document object.}

\item{file}{Character string, path to the output file (.docx).}
}
\value{
Invisibly returns the document object.
}
\description{
Saves a Word document to a file.
}
\examples{
\dontrun{
doc <- table1_word_doc() |>
  implement_table1(table1, caption = "Table 1")
save_table1_doc(doc, "output/table1.docx")
}

}
